pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-calendar-mcp-server"
version = "1.0.0"
description = "Email and Calendar Manager MCP Server for Claude integration"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your-email@example.com"}
]
license = {text = "MIT"}

dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.28.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.108.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "pydantic>=2.8.1",
    "pydantic-settings>=2.0.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
